requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.0

# LLM client — Google Gemini
google-generativeai==0.8.3
google-api-core==2.24.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1  # Required by FastAPI TestClient

# Optional: production server
# gunicorn==23.0.0

# Frontend
streamlit>=1.56.0